Validation plus Authorization Functionality Added

+++ src/Web/SportsNews.Web/Controllers/ArticlesController.cs
@@ -20,7 +20,7 @@
             this.articlesService = articlesService;
             this.categoriesService = categoriesService;
         }
-
+        [Authorize]
         public IActionResult AllByCategory(int id)
         {
             var articles = this.articlesService.GetAllByCategory(id);
@@ -31,6 +31,7 @@
             return this.View(viewModel);
         }
 
+        [Authorize]
         public IActionResult All()
         {
             var articles = this.articlesService.GetArticles();
@@ -40,64 +41,85 @@
             };
             return this.View(viewModel);
         }
-
-        [Authorize]
+        
         public IActionResult Create()
         {
-            this.ViewData["Categories"] = this.categoriesService.GetAll()
-                .Select(x => new SelectListItem
-                {
-                    Value = x.Id.ToString(),
-                    Text = x.Name
-                });
-            return this.View();
+            if (this.User.IsInRole("Administrator"))
+            {
+                this.ViewData["Categories"] = this.categoriesService.GetAll()
+                    .Select(x => new SelectListItem
+                    {
+                        Value = x.Id.ToString(),
+                        Text = x.Name
+                    });
+                return this.View();
+            }
+
+            return this.RedirectToAction("Index", "Home");
         }
 
         [HttpPost]
         public async Task<IActionResult> Create(CreateArticleInputModel input)
         {
-            if (!this.ModelState.IsValid)
+            if (this.User.IsInRole("Administrator"))
             {
-                return this.View(input);
+                if (!this.ModelState.IsValid)
+                {
+                    return this.View(input);
+                }
+
+                var id = await this.articlesService.Create(input.CategoryId, input.Content, input.Title);
+                return this.RedirectToAction("Details", new { id = id });
             }
-
-            var id = await this.articlesService.Create(input.CategoryId,input.Content, input.Title);
-            return this.RedirectToAction("Details", new {id = id});
+            return this.RedirectToAction("Index", "Home");
         }
-
-        [Authorize]
+        
         public IActionResult Update(int id)
         {
-            var model = this.articlesService.GetArticles()
-                .Where(x => x.Id == id)
-                .Select(x => new UpdateArticleInputModel
-                {
-                    Content = x.Content,
-                    CategoryId = x.CategoryId
-                }).FirstOrDefault();
-            return this.View(model);
+            if (this.User.IsInRole("Administrator"))
+            {
+                var model = this.articlesService.GetArticles()
+                    .Where(x => x.Id == id)
+                    .Select(x => new UpdateArticleInputModel
+                    {
+                        Content = x.Content,
+                        CategoryId = x.CategoryId,
+                        Title = x.Title
+                    }).FirstOrDefault();
+                return this.View(model);
+            }
+            return this.RedirectToAction("Index", "Home");
         }
 
         [HttpPost]
         public async Task<IActionResult> Update(UpdateArticleInputModel input)
         {
-            if (!this.ModelState.IsValid)
+            if (this.User.IsInRole("Administrator"))
             {
-                return this.View(input);
-            }
 
-            await this.articlesService.Update(input.Id, input.Content, input.CategoryId);
-            return this.RedirectToAction("Details", new { id = input.Id });
+                if (!this.ModelState.IsValid)
+                {
+                    return this.View(input);
+                }
+
+                await this.articlesService.Update(input.Id, input.Content, input.CategoryId, input.Title);
+                return this.RedirectToAction("Details", new { id = input.Id });
+            }
+            return this.RedirectToAction("Index", "Home");
+
         }
 
         [HttpPost]
         public async Task<IActionResult> Delete(int id)
         {
-
-            await this.articlesService.Delete(id);
+            if (this.User.IsInRole("Administrator"))
+            {
+                await this.articlesService.Delete(id);
+            }
             return this.RedirectToAction("Index", "Home");
         }
 
+        [Authorize]
         public IActionResult Details(int id)
         {
             var article = this.articlesService.GetArticleById<ArticleDetailsViewModel>(id);
